# Parsing performance improvements (not for direct merge)