--- a/Manatee.Json/Internal/SmallBufferCache.cs
+++ b/Manatee.Json/Internal/SmallBufferCache.cs
@@ -4,27 +4,27 @@ using System.Text;
 
 namespace Manatee.Json.Internal
 {
-    internal static class SmallBufferCache
-    {
-        static readonly int bufferSize = 8;
+	internal static class SmallBufferCache
+	{
+		static readonly int bufferSize = 8;
 
-        static readonly ObjectCache<char[]> cache = new ObjectCache<char[]>(() => new char[bufferSize]);
+		static readonly ObjectCache<char[]> cache = new ObjectCache<char[]>(() => new char[bufferSize]);
 
-        public static char[] Acquire(int size)
-        {
-            if (size <= bufferSize)
-                return cache.Acquire();
+		public static char[] Acquire(int size)
+		{
+			if (size <= bufferSize)
+				return cache.Acquire();
 
-            return new char[bufferSize];
-        }
+			return new char[bufferSize];
+		}
 
-        public static void Release(char[] buffer)
-        {
-            if (buffer != null)
-            {
-                if (buffer.Length <= bufferSize)
-                    cache.Release(buffer);
-            }
-        }
-    }
+		public static void Release(char[] buffer)
+		{
+			if (buffer != null)
+			{
+				if (buffer.Length <= bufferSize)
+					cache.Release(buffer);
+			}
+		}
+	}
 }