--- a/Manatee.Json/Internal/StringBuilderCache.cs
+++ b/Manatee.Json/Internal/StringBuilderCache.cs
@@ -5,27 +5,27 @@ using System.Threading;
 
 namespace Manatee.Json.Internal
 {
-    internal static class StringBuilderCache
-    {
-        private static readonly ObjectCache<StringBuilder> cache = new ObjectCache<StringBuilder>(() => new StringBuilder());
+	internal static class StringBuilderCache
+	{
+		private static readonly ObjectCache<StringBuilder> cache = new ObjectCache<StringBuilder>(() => new StringBuilder());
 
-        public static StringBuilder Acquire() => cache.Acquire();
+		public static StringBuilder Acquire() => cache.Acquire();
 
-        public static void Release(StringBuilder sb)
-        {
-            // Don't hold onto string builders that are too large
-            if (sb.Capacity < 360)
-            {
-                sb.Clear();
-                cache.Release(sb);
-            }
-        }
+		public static void Release(StringBuilder sb)
+		{
+			// Don't hold onto string builders that are too large
+			if (sb.Capacity < 360)
+			{
+				sb.Clear();
+				cache.Release(sb);
+			}
+		}
 
-        public static string GetStringAndRelease(StringBuilder sb)
-        {
-            string result = sb.ToString();
-            Release(sb);
-            return result;
-        }
-    }
+		public static string GetStringAndRelease(StringBuilder sb)
+		{
+			string result = sb.ToString();
+			Release(sb);
+			return result;
+		}
+	}
 }