--- a/Manatee.Json/Internal/StringExtensions.cs
+++ b/Manatee.Json/Internal/StringExtensions.cs
@@ -4,6 +4,7 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Manatee.Json.Internal
@@ -12,7 +13,8 @@ namespace Manatee.Json.Internal
 	{
 		private static readonly IEnumerable<char> AvailableChars = Enumerable.Range(ushort.MinValue, ushort.MaxValue)
 		                                                                     .Select(n => (char) n)
-		                                                                     .Where(c => !char.IsControl(c));
+		                                                                     .Where(c => !char.IsControl(c))
+                                                                             .ToArray();
 		private static readonly Regex _generalEscapePattern = new Regex("%(?<Value>[0-9A-F]{2})", RegexOptions.IgnoreCase);
 
 		public static string EvaluateEscapeSequences(this string source, out string result)
@@ -117,6 +119,7 @@ namespace Manatee.Json.Internal
 			}
 			return source;
 		}
+
 		private static string _Replace(string source, int index, int count, string content)
 		{
 			// I've checked both of these methods with ILSpy.  They occur in external methods, so
@@ -125,75 +128,59 @@ namespace Manatee.Json.Internal
 		}
 		public static string SkipWhiteSpace(this string source, ref int index, int length, out char ch)
 		{
-			if (index >= length)
-			{
-				ch = default(char);
-				return "Unexpected end of input.";
-			}
-			var c = source[index];
+            ch = default(char);
 			while (index < length)
 			{
-				if (!char.IsWhiteSpace(c)) break;
+                ch = source[index];
+				if (!char.IsWhiteSpace(ch)) break;
 				index++;
-				if (index >= length)
-				{
-					ch = default(char);
-					return "Unexpected end of input.";
-				}
-				c = source[index];
 			}
+
 			if (index >= length)
 			{
 				ch = default(char);
 				return "Unexpected end of input.";
 			}
-			ch = c;
+
 			return null;
 		}
-		public static string SkipWhiteSpace(this StreamReader stream, out char ch)
+		public static string SkipWhiteSpace(this TextReader stream, out char ch)
 		{
-			if (stream.EndOfStream)
-			{
-				ch = default(char);
-				return "Unexpected end of input.";
-			}
 			ch = (char) stream.Peek();
-			while (!stream.EndOfStream)
+			while (ch != -1)
 			{
 				if (!char.IsWhiteSpace(ch)) break;
 				stream.Read();
 				ch = (char) stream.Peek();
 			}
-			if (stream.EndOfStream)
+
+			if (ch == -1)
 			{
 				ch = default(char);
 				return "Unexpected end of input.";
 			}
+
 			return null;
 		}
-		public static async Task<(string, char)> SkipWhiteSpaceAsync(this StreamReader stream)
-		{
-			char ch;
-			if (stream.EndOfStream)
-			{
-				ch = default(char);
-				return ("Unexpected end of input.", ch);
-			}
-			ch = (char) stream.Peek();
-			while (!stream.EndOfStream)
-			{
-				if (!char.IsWhiteSpace(ch)) break;
-				await stream.TryRead();
-				ch = (char) stream.Peek();
-			}
-			if (stream.EndOfStream)
-			{
-				ch = default(char);
-				return ("Unexpected end of input.", ch);
-			}
-			return (null, ch);
-		}
-		public static string UnescapePointer(this string reference)
+        public static async Task<(string, char)> SkipWhiteSpaceAsync(this TextReader stream, char[] scratch)
+        {
+            System.Diagnostics.Debug.Assert(scratch.Length >= 1);
+            char ch;
+            ch = (char)stream.Peek();
+            while (ch != -1)
+            {
+                if (!char.IsWhiteSpace(ch)) break;
+                await stream.ReadAsync(scratch, 0, 1);
+                ch = (char)stream.Peek();
+            }
+            if (ch == -1)
+            {
+                ch = default(char);
+                return ("Unexpected end of input.", ch);
+            }
+            return (null, ch);
+        }
+        public static string UnescapePointer(this string reference)
 		{
 			var unescaped = reference.Replace("~1", "/")
 			                         .Replace("~0", "~");
@@ -206,12 +193,18 @@ namespace Manatee.Json.Internal
 			}
 			return unescaped;
 		}
+        public static Task<bool> TryRead(this TextReader stream, char[] buffer, int offset, int count)
+        {
+            return TryRead(stream, buffer, offset, count, CancellationToken.None);
+        }
+        public static async Task<bool> TryRead(this TextReader stream, char[] buffer, int offset, int count, CancellationToken token)
+        {
+            if (token.IsCancellationRequested)
+                return false;
 
-		public static async Task<(bool success, char c)> TryRead(this StreamReader stream)
-		{
-			var buffer = new char[1];
-			var count = await stream.ReadAsync(buffer, 0, 1);
-			return (count != 0, buffer[0]);
-		}
-	}
+            var charsRead = await stream.ReadBlockAsync(buffer, offset, count);
+
+            return count == charsRead;
+        }
+    }
 }