--- a/Manatee.Json/JsonValue.cs
+++ b/Manatee.Json/JsonValue.cs
@@ -46,8 +46,8 @@ namespace Manatee.Json
 			{
 				if (Type != JsonValueType.Boolean)
 					return !JsonOptions.ThrowOnIncorrectTypeAccess
-						       ? default(bool)
-						       : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Boolean);
+							   ? default(bool)
+							   : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Boolean);
 				return _boolValue;
 			}
 			private set
@@ -72,8 +72,8 @@ namespace Manatee.Json
 			{
 				if (Type != JsonValueType.String)
 					return !JsonOptions.ThrowOnIncorrectTypeAccess
-						       ? default(string)
-						       : throw new JsonValueIncorrectTypeException(Type, JsonValueType.String);
+							   ? default(string)
+							   : throw new JsonValueIncorrectTypeException(Type, JsonValueType.String);
 				return _stringValue;
 			}
 			private set
@@ -98,8 +98,8 @@ namespace Manatee.Json
 			{
 				if (Type != JsonValueType.Number)
 					return !JsonOptions.ThrowOnIncorrectTypeAccess
-						       ? default(double)
-						       : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Number);
+							   ? default(double)
+							   : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Number);
 				return _numberValue;
 			}
 			private set
@@ -124,8 +124,8 @@ namespace Manatee.Json
 			{
 				if (Type != JsonValueType.Object)
 					return !JsonOptions.ThrowOnIncorrectTypeAccess
-						       ? default(JsonObject)
-						       : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Object);
+							   ? default(JsonObject)
+							   : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Object);
 				return _objectValue;
 			}
 			private set
@@ -150,8 +150,8 @@ namespace Manatee.Json
 			{
 				if (Type != JsonValueType.Array)
 					return !JsonOptions.ThrowOnIncorrectTypeAccess
-						       ? default(JsonArray)
-						       : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Array);
+							   ? default(JsonArray)
+							   : throw new JsonValueIncorrectTypeException(Type, JsonValueType.Array);
 				return _arrayValue;
 			}
 			private set