This PR is for review/comments only and should probably be brought in in parts, depending on the goals of the project (readability vs performance).

General strategy:
1. Reduce allocations.
2. Pool allocations.
3. Reduce LINQ usage.

BenchmarkDotNet results for the PR compared to 9.5.X and Newtonsoft.Json 10.X (read the commit messages for their individual improvements):
```
BenchmarkDotNet=v0.10.11, OS=Windows 7 SP1 (6.1.7601.0)
Processor=Intel Xeon CPU E5-2620 v3 2.40GHzIntel Xeon CPU E5-2620 v3 2.40GHz, ProcessorCount=24
Frequency=2338369 Hz, Resolution=427.6485 ns, Timer=TSC
  [Host]     : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2114.0
  DefaultJob : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2114.0

                            Method |     Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
---------------------------------- |---------:|----------:|----------:|----------:|---------:|---------:|----------:|
 JsonSerializer_Deserialize_Stream | 38.16 ms | 0.9832 ms | 2.8990 ms | 1062.5000 | 437.5000 | 125.0000 |   5.76 MB |
                     JObject_Parse | 28.57 ms | 0.8978 ms | 2.6470 ms | 1180.0000 | 495.9375 | 159.6875 |   6.26 MB |
                     
            JsonValue_Parse_Stream | 54.46 ms | 1.7220 ms | 5.0770 ms | 4062.5000 | 625.0000 | 187.5000 |  23.54 MB |
            JsonValue_Parse_String | 42.52 ms | 1.0770 ms | 3.1740 ms | 4062.5000 | 625.0000 | 187.5000 |  23.54 MB |
              JsonValue_ParseAsync | 66.57 ms | 1.3300 ms | 3.2630 ms | 4062.5000 | 625.0000 | 187.5000 |  23.55 MB |
              
        New_JsonValue_Parse_Stream | 27.34 ms | 0.9102 ms | 2.6836 ms |  625.0000 | 312.5000 |        - |   3.96 MB |
        New_JsonValue_Parse_String | 12.49 ms | 0.2479 ms | 0.2198 ms |  687.5000 | 312.5000 |  62.5000 |   3.95 MB |
          New_JsonValue_ParseAsync | 28.94 ms | 0.7739 ms | 2.2697 ms |  656.2500 | 312.5000 |  62.5000 |   3.96 MB |
```

I also need to fix-up my whitespace to match yours.