@gregsdennis 2017-12-07
Overall, I like what you've done.  I'm impressed with the improvements around reorganizing the individual parser logic, and I like the StringBuilder and buffer caches.

I understand the need for some of the changes, like removing `.In()` and using switches instead.  It's not as pretty, but I think not having to convert parameters into an array then iterating that array is probably more important.

I'd like to pull this into a branch and make some stylistic changes, but I think the functional part is good.