@sixlettervariables 2017-12-07
Sounds good. The caching comes from [Roslyn](http://mattwarren.org/2014/06/10/roslyn-code-base-performance-lessons-part-2/), and should be cleaned up for inclusion. I'm not sure all of the defaults are appropriate.

I'm now taking a look at the serialization/deserialization to object side.