@gregsdennis 2017-12-08
@sixlettervariables I appreciate the effort.  Serialization has become something of a rat's nest of attempts to cache reflection and architecture out the different serialization methods.  That's why I haven't gone back in to tackle it yet.