@gregsdennis 2017-12-08
@sixlettervariables I see 13 parsing errors in the tests.  Most of them have to do with unicode escape sequences, but a few deal with parsing lone explicit values (`null`, `true`, & `false`).