@gregsdennis 2017-12-08
@sixlettervariables please pull my changes from https://github.com/gregsdennis/Manatee.Json/tree/task/issue-9-parsing-perf-improvements.  I have all of the tests passing.